+++ crates/weaver_semconv/src/registry.rs
@@ -3,7 +3,7 @@
 //! Semantic Convention Registry.
 
 use crate::attribute::AttributeSpecWithProvenance;
-use crate::group::GroupSpecWithProvenance;
+use crate::group::{GroupSpecWithProvenance, ImportsWithProvenance};
 use crate::json_schema::JsonSchemaValidator;
 use crate::manifest::RegistryManifest;
 use crate::metric::MetricSpecWithProvenance;
@@ -246,6 +246,19 @@ impl SemConvRegistry {
             })
     }
 
+    /// Returns an iterator over all the unresolved imports defined in the semantic convention
+    /// registry. Each import is associated with its provenance (path or URL).
+    pub fn unresolved_imports_iter(&self) -> impl Iterator<Item = ImportsWithProvenance> + '_ {
+        self.specs
+            .iter()
+            .flat_map(|SemConvSpecWithProvenance { spec, provenance }| {
+                spec.imports.iter().map(|imports| ImportsWithProvenance {
+                    imports: imports.clone(),
+                    provenance: provenance.clone(),
+                })
+            })
+    }
+
     /// Returns a set of stats about the semantic convention registry.
     pub fn stats(&self) -> Stats {
         Stats {
@@ -348,6 +361,7 @@ mod tests {
                         entity_associations: Vec::new(),
                         value_type: None,
                     }],
+                    imports: None,
                 },
             ),
             (
@@ -375,6 +389,7 @@ mod tests {
                         entity_associations: Vec::new(),
                         value_type: None,
                     }],
+                    imports: None,
                 },
             ),
         ];